requirements.txt
azure-cognitiveservices-speech
rich
openai